[deps] remove extra includes; fix linking problems

+++ Glance/Glance/Common/OpenGLPrimitive/Curves/CubicBezier/CubicBezier.cpp
@@ -1,7 +1,4 @@
 #include "../../../../pch.h"
-#include <glm/glm.hpp>
-#include <glm/gtx/rotate_vector.hpp>
-#include <glm/gtc/type_ptr.hpp>
 #include "CubicBezier.h"
 
 CubicBezier::CubicBezier(
@@ -14,7 +11,7 @@ CubicBezier::CubicBezier(
 
 	std::vector<RGB> finishColors = GetVerticesColor(colors, m_verticesNumber + 1);
 
-	for (int i = 0; i <= m_verticesNumber; i++)
+	for (int i = 0; i <= static_cast<int>(m_verticesNumber); i++)
 	{
 		GLfloat t = GLfloat(i) / GLfloat(m_verticesNumber);
 		GLfloat x = static_cast<float>(pow(1 - t, 3) * controlPoints.p1.x + 3 * t * pow(1 - t, 2) * controlPoints.p2.x + 3 * pow(t, 2) * (1 - t) * controlPoints.p3.x + pow(t, 3) * controlPoints.p4.x);
